@@ -89,20 +89,21 @@
     "bittorrent-tracker": "9.19.0",
     "browserify": "17.0.0",
     "disc": "1.3.3",
+    "esmify": "^2.1.1",
     "finalhandler": "1.2.0",
     "network-address": "1.1.2",
     "run-series": "1.1.9",
-    "semantic-release": "19.0.3",
+    "semantic-release": "19.0.5",
     "serve-static": "1.15.0",
     "standard": "*",
-    "tape": "5.5.3",
+    "tape": "5.6.1",
     "webtorrent-fixtures": "1.7.5"
   },
   "optionalDependencies": {
     "utp-native": "^2.5.3"
   },
   "engines": {
-    "node": ">=12"
+    "node": ">=14"
   },
   "funding": [
     {
