@@ -1,15 +1,12 @@
-const EventEmitter = require('events')
-const { PassThrough } = require('stream')
-const streamToBlob = require('stream-to-blob')
-const streamToBlobURL = require('stream-to-blob-url')
-const streamToBuffer = require('stream-with-known-length-to-buffer')
-const queueMicrotask = require('queue-microtask')
-const rangeParser = require('range-parser')
-const mime = require('mime')
-const eos = require('end-of-stream')
-const FileStream = require('./file-stream.js')
+import EventEmitter from 'events'
+import { PassThrough } from 'streamx'
+import { BlobWriteStream } from 'fast-blob-stream'
+import streamToBuffer from 'stream-with-known-length-to-buffer'
+import queueMicrotask from 'queue-microtask'
+import mime from 'mime'
+import FileStream from './file-stream.js'
 
-class File extends EventEmitter {
+export default class File extends EventEmitter {
   constructor (torrent, file) {
     super()
 
@@ -121,73 +118,16 @@ class File extends EventEmitter {
 
   getBlob (cb) {
     if (typeof window === 'undefined') throw new Error('browser-only method')
-    streamToBlob(this.createReadStream(), mime.getType(this.name))
-      .then(
-        blob => cb(null, blob),
-        err => cb(err)
-      )
+    const writeStream = new BlobWriteStream(blob => {
+      cb(null, blob)
+    }, { mimeType: mime.getType(this.name) || 'application/octet-stream' })
+    this.createReadStream().pipe(writeStream)
   }
 
   getBlobURL (cb) {
-    if (typeof window === 'undefined') throw new Error('browser-only method')
-    streamToBlobURL(this.createReadStream(), mime.getType(this.name))
-      .then(
-        blobUrl => cb(null, blobUrl),
-        err => cb(err)
-      )
-  }
-
-  _serve (req) {
-    const res = {
-      status: 200,
-      headers: {
-        // Support range-requests
-        'Accept-Ranges': 'bytes',
-        'Content-Type': mime.getType(this.name),
-        'Cache-Control': 'no-cache, no-store, must-revalidate, max-age=0',
-        Expires: '0'
-      },
-      body: req.method === 'HEAD' ? '' : 'STREAM'
-    }
-    // force the browser to download the file if if it's opened in a new tab
-    if (req.destination === 'document') {
-      res.headers['Content-Type'] = 'application/octet-stream'
-      res.headers['Content-Disposition'] = 'attachment'
-      res.body = 'DOWNLOAD'
-    }
-
-    // `rangeParser` returns an array of ranges, or an error code (number) if
-    // there was an error parsing the range.
-    let range = rangeParser(this.length, req.headers.range || '')
-
-    if (range.constructor === Array) {
-      res.status = 206 // indicates that range-request was understood
-
-      // no support for multi-range request, just use the first range
-      range = range[0]
-
-      res.headers['Content-Range'] = `bytes ${range.start}-${range.end}/${this.length}`
-      res.headers['Content-Length'] = `${range.end - range.start + 1}`
-    } else {
-      res.headers['Content-Length'] = this.length
-    }
-
-    const stream = req.method === 'GET' && this.createReadStream(range)
-
-    let pipe = null
-    if (stream) {
-      this.emit('stream', { stream, req, file: this }, piped => {
-        pipe = piped
-
-        // piped stream might not close the original filestream on close/error, this is agressive but necessary
-        eos(piped, () => {
-          if (piped) piped.destroy()
-          stream.destroy()
-        })
-      })
-    }
-
-    return [res, pipe || stream, pipe && stream]
+    this.getBlob((_err, blob) => {
+      cb(null, URL.createObjectURL(blob))
+    })
   }
 
   getStreamURL () {
@@ -215,5 +155,3 @@ class File extends EventEmitter {
     this._fileStreams.clear()
   }
 }
-
-module.exports = File
