@@ -1,10 +1,10 @@
-const http = require('http')
-const escapeHtml = require('escape-html')
-const mime = require('mime')
-const pump = require('pump')
-const rangeParser = require('range-parser')
-const queueMicrotask = require('queue-microtask')
-const { Readable } = require('stream')
+import http from 'http'
+import escapeHtml from 'escape-html'
+import mime from 'mime'
+import pump from 'pump'
+import rangeParser from 'range-parser'
+import queueMicrotask from 'queue-microtask'
+import { Readable } from 'streamx'
 
 const keepAliveTime = 20000
 
@@ -423,4 +423,4 @@ function encodeRFC5987 (str) {
     .replace(/%(?:7C|60|5E)/g, unescape)
 }
 
-module.exports = { NodeServer, BrowserServer }
+export { NodeServer, BrowserServer }
