@@ -1,41 +1,44 @@
 /* global Blob */
 
-const EventEmitter = require('events')
-const fs = require('fs')
-const net = require('net') // browser exclude
-const os = require('os') // browser exclude
-const path = require('path')
-const addrToIPPort = require('addr-to-ip-port')
-const { default: BitField } = require('bitfield')
-const CacheChunkStore = require('cache-chunk-store')
-const ChunkStoreWriteStream = require('chunk-store-stream/write')
-const cpus = require('cpus')
-const debugFactory = require('debug')
-const Discovery = require('torrent-discovery')
-const FSChunkStore = require('fs-chunk-store') // browser: `memory-chunk-store`
-const get = require('simple-get')
-const ImmediateChunkStore = require('immediate-chunk-store')
-const ltDontHave = require('lt_donthave')
-const MemoryChunkStore = require('memory-chunk-store')
-const HybridChunkStore = require('hybrid-chunk-store')
-const MultiStream = require('multistream')
-const parallel = require('run-parallel')
-const parallelLimit = require('run-parallel-limit')
-const parseTorrent = require('parse-torrent')
-const Piece = require('torrent-piece')
-const pump = require('pump')
-const queueMicrotask = require('queue-microtask')
-const randomIterate = require('random-iterate')
-const sha1 = require('simple-sha1')
-const throughput = require('throughput')
-const utMetadata = require('ut_metadata')
-const utPex = require('ut_pex') // browser exclude
+import EventEmitter from 'events'
+import fs from 'fs'
+import net from 'net' // browser exclude
+import os from 'os' // browser exclude
+import path from 'path'
+import addrToIPPort from 'addr-to-ip-port'
+import BitField from 'bitfield'
+import CacheChunkStore from 'cache-chunk-store'
+import ChunkStoreWriteStream from 'chunk-store-stream/write.js'
+import cpus from 'cpus'
+import debugFactory from 'debug'
+import Discovery from 'torrent-discovery'
+import FSChunkStore from 'fs-chunk-store' // browser: `memory-chunk-store`
+import get from 'simple-get'
+import ImmediateChunkStore from 'immediate-chunk-store'
+import ltDontHave from 'lt_donthave'
+import MemoryChunkStore from 'memory-chunk-store'
+import HybridChunkStore from 'hybrid-chunk-store'
+import joinIterator from 'join-async-iterator'
+import parallel from 'run-parallel'
+import parallelLimit from 'run-parallel-limit'
+import parseTorrent from 'parse-torrent'
+import Piece from 'torrent-piece'
+import pump from 'pump'
+import queueMicrotask from 'queue-microtask'
+import randomIterate from 'random-iterate'
+import sha1 from 'simple-sha1'
+import throughput from 'throughput'
+import utMetadata from 'ut_metadata'
+import utPex from 'ut_pex' // browser exclude
+import { Readable } from 'streamx'
 
-const File = require('./file.js')
-const Peer = require('./peer.js')
-const RarityMap = require('./rarity-map.js')
-const utp = require('./utp.js') // browser exclude
-const WebConn = require('./webconn.js')
+import File from './file.js'
+import Peer from './peer.js'
+import RarityMap from './rarity-map.js'
+import utp from './utp.js' // browser exclude
+import WebConn from './webconn.js'
+
+import info from '../package.json' assert { type: 'json' }
 
 const debug = debugFactory('webtorrent:torrent')
 const MAX_BLOCK_LENGTH = 128 * 1024
@@ -56,7 +59,7 @@ const FILESYSTEM_CONCURRENCY = process.browser ? cpus().length : 2
 
 const RECONNECT_WAIT = [1_000, 5_000, 15_000]
 
-const VERSION = require('../package.json').version
+const VERSION = info.version
 const USER_AGENT = `WebTorrent/${VERSION} (https://webtorrent.io)`
 
 let TMP
@@ -66,7 +69,7 @@ try {
   TMP = path.join(typeof os.tmpdir === 'function' ? os.tmpdir() : '/', 'webtorrent')
 }
 
-class Torrent extends EventEmitter {
+export default class Torrent extends EventEmitter {
   constructor (torrentId, client, opts) {
     super()
 
@@ -1841,7 +1844,7 @@ class Torrent extends EventEmitter {
     if (!Array.isArray(streams)) streams = [streams]
     if (!cb) cb = noop
 
-    const readable = new MultiStream(streams)
+    const readable = Readable.from(joinIterator(streams))
     const writable = new ChunkStoreWriteStream(this.store, this.pieceLength)
 
     pump(readable, writable, err => {
@@ -1997,5 +2000,3 @@ function randomInt (high) {
 }
 
 function noop () {}
-
-module.exports = Torrent
