@@ -1,6 +1,6 @@
-module.exports = (() => {
+export default (() => {
   try {
-    return require('utp-native')
+    return import('utp-native')
   } catch (err) {
     console.warn('WebTorrent: uTP not supported')
     return {}
