@@ -1,6 +1,6 @@
 /* eslint-env serviceworker */
 
-const fileResponse = require('./worker-server.js')
+import fileResponse from './worker-server.js'
 
 self.addEventListener('install', () => {
   self.skipWaiting()
