@@ -1,8 +1,8 @@
-const fixtures = require('webtorrent-fixtures')
-const test = require('tape')
-const MemoryChunkStore = require('memory-chunk-store')
-const randombytes = require('randombytes')
-const WebTorrent = require('../../index.js')
+import fixtures from 'webtorrent-fixtures'
+import test from 'tape'
+import MemoryChunkStore from 'memory-chunk-store'
+import randombytes from 'randombytes'
+import WebTorrent from '../../index.js'
 
 test('client.add: emit torrent events in order', t => {
   t.plan(6)
