@@ -4,7 +4,7 @@
 const portTimeoutDuration = 5000
 let cancellable = false
 
-module.exports = event => {
+const listener = event => {
   const { url } = event.request
   if (!url.includes(self.registration.scope + 'webtorrent/')) return null
   if (url.includes(self.registration.scope + 'webtorrent/keepalive/')) return new Response()
@@ -17,6 +17,8 @@ module.exports = event => {
   }
   return serve(event)
 }
+
+export default listener
 
 async function serve ({ request }) {
   const { url, method, headers, destination } = request
